Updated Repositories, and added Thread Hits feat

+++ app/database/migrations/2014_02_05_122231_add_hits_to_threads_table.php
@@ -13,7 +13,7 @@ class AddHitsToThreadsTable extends Migration {
 	public function up()
 	{
 		Schema::table('threads', function($table) {
-			$table->bigInteger('hits');
+			$table->integer('hits');
 		});
 	}
